Made changes in /prediction def.
Made a debugger inside the def predict to solve the issue of the model receiving 58 features instead of 57. This debugger implements a dropping function for the model to successfully obtain 57 features and solve the problem of the prediction.html page not loading to show the results. Further changes have to be made after deciding what features to include in the form since this forcefully drops columns that may seem important for the model.